feat: reafctor subgraph, lineage graph, children at stop layer support
