+++ pychunkedgraph/app/segmentation/common.py
@@ -578,6 +578,7 @@ def handle_leaves(table_id, root_id):
     user_id = str(g.auth_user["id"])
     current_app.user_id = user_id
 
+    stop_layer = int(request.args.get("stop_layer", 1))
     if "bounds" in request.args:
         bounds = request.args["bounds"]
         bounding_box = np.array(
@@ -586,13 +587,19 @@ def handle_leaves(table_id, root_id):
     else:
         bounding_box = None
 
-    # Call ChunkedGraph
     cg = app_utils.get_cg(table_id)
-    atomic_ids = cg.get_subgraph(
-        int(root_id), bbox=bounding_box, bbox_is_coordinate=True, leaves_only=True
+    if stop_layer > 1:
+        return cg.get_subgraph_nodes(
+            int(root_id),
+            bbox=bounding_box,
+            bbox_is_coordinate=True,
+            return_layers=[stop_layer]
+        )
+    return cg.get_subgraph_leaves(
+        int(root_id),
+        bbox=bounding_box,
+        bbox_is_coordinate=True,
     )
-
-    return atomic_ids
 
 
 ### LEAVES OF MANY ROOTS ---------------------------------------------------------------------
